Move examples to separate files - third part
